pyproject.toml
[tool.poetry]
name = "drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
pillow = "^10.2.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
django-filter = "^23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
